♻️refactor: 별점을 double로 변경
- 별점은 0.5단위로 매길 수 있다

+++ src/main/java/checkmo/domain/club/entity/BookRecommend.java
@@ -18,7 +18,7 @@ public class BookRecommend extends BaseEntity {
 
     private String content;
 
-    private Integer rate;
+    private double rate;
 
     private boolean isCheck;
